+++ ksyun/config.go
@@ -15,6 +15,7 @@ import (
 	"github.com/KscSDK/ksc-sdk-go/service/kcsv1"
 	"github.com/KscSDK/ksc-sdk-go/service/kcsv2"
 	"github.com/KscSDK/ksc-sdk-go/service/kec"
+	"github.com/KscSDK/ksc-sdk-go/service/knad"
 	"github.com/KscSDK/ksc-sdk-go/service/krds"
 	"github.com/KscSDK/ksc-sdk-go/service/mongodb"
 	"github.com/KscSDK/ksc-sdk-go/service/rabbitmq"
@@ -81,6 +82,7 @@ func (c *Config) Client() (*KsyunClient, error) {
 	client.tagv1conn = tag.SdkNew(cli, cfg, url)
 	client.kceconn = kce.SdkNew(cli, cfg, url)
 	client.kcev2conn = kcev2.SdkNew(cli, cfg, url)
+	client.knadconn = knad.SdkNew(cli, cfg, url)
 
 	//懒加载ks3-client 所以不在此初始化
 	return &client, nil